CHG: add a tokenizer and refine the dataset
