+++ nlp_practice/case/translation/data/dataset.py
@@ -1,0 +1,47 @@
+class PairDataset:
+    """
+    Examples
+    --------
+    >>> from nlp_practice.case.translation.data.dataset import PairDataset
+    >>> from nlp_practice.case.translation.data.preprocessor import Preprocessor
+    >>> input_language, output_language, pairs = Preprocessor(
+    ...     base_path="examples/translation/data",
+    ...     first_language="eng",
+    ...     second_language="fra",
+    ...     does_reverse=True,
+    ... ).process()
+    >>> dataset = PairDataset(pairs, input_language, output_language)()
+    >>> len(dataset)
+    11445
+    >>> type(dataset())
+    <torch.utils.data.dataset.TensorDataset at 0x117961390>
+    """
+
+    def __init__(
+        self,
+        pairs: list[list[str]],
+        input_language: LanguageData,
+        output_language: LanguageData,
+    ) -> None:
+        self.pairs = pairs
+        self.input_language = input_language
+        self.output_language = output_language
+
+    def __len__(self) -> TensorDataset:
+        return len(self.pairs)
+
+    def __call__(self) -> TensorDataset:
+        n = len(self.pairs)
+        input_ids = np.zeros((n, MAX_LENGTH), dtype=np.int32)
+        target_ids = np.zeros((n, MAX_LENGTH), dtype=np.int32)
+
+        for index, (input, target) in enumerate(self.pairs):
+            inputs = index_from_sentence(self.input_language, input)
+            targets = index_from_sentence(self.output_language, target)
+
+            inputs.append(EOS_TOKEN)
+            targets.append(EOS_TOKEN)
+            input_ids[index, : len(inputs)] = inputs
+            target_ids[index, : len(targets)] = targets
+
+        return TensorDataset(torch.LongTensor(input_ids), torch.LongTensor(target_ids))
